Fixed issue with priority queue
Encoding now returns a string, as well as the huffman tree (not sure if this should though). Making the tree crashes at the end, but the professor told me that this approach is likely the best way of going about it. Need help with fixing this, the rest of the program should work fine.